User and Tag Policies

+++ app/Http/Controllers/TagController.php
@@ -82,6 +82,8 @@ class TagController extends Controller
      */
     public function edit(Tag $tag)
     {
+        abort_unless(Gate::allows('update', $tag), 403); // Không thể update tag nếu không phải admin
+
         return view('tag.edit')->with([
             'tag' => $tag
         ]);
@@ -96,6 +98,9 @@ class TagController extends Controller
      */
     public function update(Request $request, Tag $tag)
     {
+
+        abort_unless(Gate::allows('update', $tag), 403); // Không thể update tag nếu không phải admin
+
         $request->validate([
             'name' => 'required',
             'style' => 'required',
@@ -122,6 +127,8 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        abort_unless(Gate::allows('delete', $tag), 403); // Không thể xoá tag nếu không phải admin
+
         $oldName = $tag->name;
         $tag->delete();
         return $this->index()->with([
